chore(folder-structure): Move content app store logic to content/store folder
