+++ src/app/content/App/Contribute/PreviewScreen/withConnect.ts
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { goBack } from 'connected-react-router';
-import { ContentState } from 'app/content/store';
-import { getContribution } from 'app/content/selectors';
+import { getContribution } from 'app/content/store/selectors';
+import { ContentState } from 'app/content/store/reducers';
 import { close } from 'libs/store/actions/ui';
 import { submitContribution } from 'libs/store/actions/contribution';
