+++ src/app/content/App/Notice/List/withConnect.ts
@@ -1,12 +1,12 @@
 import { connect } from 'react-redux';
+import { getNoticesToDisplay } from 'app/content/store/selectors';
 import {
   dismissNotice as dismiss,
   undismissNotice as undismiss,
   confirmDismissNotice as confirmDismiss
 } from 'libs/store/actions/notices';
-import onContributorClick from 'app/content/actions/goToContributor';
-import { getNoticesToDisplay } from 'app/content/selectors';
-import { ContentState } from 'app/content/store';
+import onContributorClick from 'app/content/store/actions/goToContributor';
+import { ContentState } from 'app/content/store/reducers';
 
 const mapStateToProps = (state: ContentState) => ({
   notices: getNoticesToDisplay(state)
