+++ src/app/content/App/UI.tsx
@@ -11,14 +11,14 @@ import {
 import { close } from 'libs/store/actions';
 import { CloseCause } from 'libs/lmem/ui';
 import Notification from 'components/organisms/Notification';
-import { ContentState } from '../store';
+import { ContentState } from '../store/reducers';
 import {
   getTitle,
   hasServiceMessage,
   isLoaded,
   isNoticeContext,
   isOpen
-} from '../selectors';
+} from '../store/selectors';
 import Notice from './Notice';
 import Subscriptions from './Subscriptions';
 import Help from './Help';
