+++ src/app/content/store/actions/index.ts
@@ -1,8 +1,8 @@
 import {
   RemoveUITitleAction,
   SetUITitleAction
-} from 'app/content/actions/ui/title';
-import { LoadedAction } from 'app/content/actions/ui/open.actions';
+} from 'app/content/store/actions/ui/title';
+import { LoadedAction } from 'app/content/store/actions/ui/open.actions';
 import { AppAction } from 'libs/store/actions';
 
 export type ContentAction =
