+++ src/app/content/store/actions/ui/open.actions.ts
