+++ src/app/content/store/sagas/ui/fakeLoading.saga.ts
@@ -1,6 +1,6 @@
 import { delay, put } from '@redux-saga/core/effects';
 import { OpenedAction, OpenFrom } from 'libs/store/actions';
-import { loaded } from 'app/content/actions/ui/open.actions';
+import { loaded } from 'app/content/store/actions/ui/open.actions';
 
 const DELAY_BEFORE_SHOWING = process.env.NODE_ENV === 'production' ? 1100 : 110;
